Convert to 3D environment
We are moving the game from a 2D environment to a 3D one, this will allow us to easily add depth to the game, as we will be able to represent layers significantly easier. We will also be able to model structures and blocks in 3D so they can be viewed from all angles, instead of drawing 3D art in 2D for each case. All legacy 2D work has been moved to Assets/2D, and all new 3D work is in Assets/3D.